+++ build_boxes.sh
@@ -1,25 +1,37 @@
 #!/bin/bash -eux
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.45 \
+  --atlas_version=1.0.47 \
   --os=debian \
-  --os_version=8.7.1 \
-  --providers=virtualbox,vmware
+  --os_version=9.1.0 \
+  --providers=vmware,virtualbox
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.45 \
+  --atlas_version=1.0.47 \
+  --os=debian \
+  --os_version=8.9.0 \
+  --providers=vmware,virtualbox
+
+bundle exec thor packer:build \
+  --atlas_version=1.0.47 \
   --os=ubuntu \
   --os_version=14.04.4-server \
-  --providers=virtualbox,vmware
+  --providers=vmware,virtualbox
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.45 \
+  --atlas_version=1.0.47 \
   --os=ubuntu \
-  --os_version=16.04.2-server \
-  --providers=virtualbox,vmware
+  --os_version=16.04.3-server \
+  --providers=vmware,virtualbox
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.45 \
+  --atlas_version=1.0.47 \
+  --os=ubuntu \
+  --os_version=17.04-server \
+  --providers=vmware,virtualbox
+
+bundle exec thor packer:build \
+  --atlas_version=1.0.47 \
   --os=ubuntu \
   --os_version=16.10-server \
-  --providers=virtualbox,vmware
+  --providers=vmware,virtualbox